requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
google-auth
python-dotenv
psycopg2-binary
gunicorn